% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicatvarsum.R
\name{bicatvarsum}
\alias{bicatvarsum}
\title{provides a nice summary and test for a categorical variable by a grouping variable, with optional Fisher's exact test}
\usage{
bicatvarsum(
  mat,
  idcol,
  varcol,
  groupcol,
  doplot = T,
  myrot = F,
  mytest = T,
  check5 = F,
  print5 = T,
  usecolours = NULL
)
}
\description{
provides a nice summary and test for a categorical variable by a grouping variable, with optional Fisher's exact test
}
