% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicontcont.R
\name{bicontcont}
\alias{bicontcont}
\title{provides a nice summary for two continous variables with a scatterplot with lowess and the correlation}
\usage{
bicontcont(mat, idcol, xcol, ycol, check5 = F, print5 = T)
}
\description{
provides a nice summary for two continous variables with a scatterplot with lowess and the correlation
}
