% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicontvarsum.R
\name{bicontvarsum}
\alias{bicontvarsum}
\title{provides a nice summary and test for a continuous variable by a grouping variable, with optional testing among groups}
\usage{
bicontvarsum(
  mat,
  idcol,
  varcol,
  groupcol,
  myrot = F,
  mytest = T,
  myuni = T,
  check5 = F,
  print5 = T,
  usecolours = NULL
)
}
\description{
provides a nice summary and test for a continuous variable by a grouping variable, with optional testing among groups
}
