% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCG_catcontvarsum.r
\name{BioCG_catcontvarsum}
\alias{BioCG_catcontvarsum}
\title{summarize cat and cont variable}
\usage{
BioCG_catcontvarsum(data_mat, var_name, cat_var = TRUE)
}
\description{
summarize cat and cont variable
}
