% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCG_catvarsum.R
\name{BioCG_catvarsum}
\alias{BioCG_catvarsum}
\title{provides a nice summary for a categorical variable, with optional barchart}
\usage{
BioCG_catvarsum(
  mat,
  idcol,
  varcol,
  doplot = T,
  myrot = F,
  myuni = T,
  check5 = F,
  print5 = T,
  usecolours = NULL
)
}
\description{
provides a nice summary for a categorical variable, with optional barchart
}
