% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCG_contvarsum.R
\name{BioCG_contvarsum}
\alias{BioCG_contvarsum}
\title{provides a nice summary for a continuous variable, with optional boxplot, histogram}
\usage{
BioCG_contvarsum(
  mat,
  idcol,
  varcol,
  doplot = T,
  myuni = T,
  check5 = F,
  print5 = T
)
}
\description{
provides a nice summary for a continuous variable, with optional boxplot, histogram
}
