% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catkabelmat.R
\name{catkabelmat}
\alias{catkabelmat}
\title{Provides a nice kabel table to summarize a categorical variable supplied in y}
\usage{
catkabelmat(y, mylab, decimals = 1, check5 = F, print5 = T)
}
\description{
Provides a nice kabel table to summarize a categorical variable supplied in y
}
