% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catkabelmat2.R
\name{catkabelmat2}
\alias{catkabelmat2}
\title{Not needed likely as the calcs aren't so ownerous that we cant just use catkabelmat}
\usage{
catkabelmat2(ytable, mylab, decimals = 1)
}
\description{
Not needed likely as the calcs aren't so ownerous that we cant just use catkabelmat
}
