% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cattable.R
\name{cattable}
\alias{cattable}
\title{returns a data.frame of counts and percents corresponding to the data y that is good for exporting or printing with knitr::kable}
\usage{
cattable(y, mydigits = 1, check5 = F, print5 = T)
}
\description{
returns a data.frame of counts and percents corresponding to the data y that is good for exporting or printing with knitr::kable
}
