% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contkabelmat.R
\name{contkabelmat}
\alias{contkabelmat}
\title{Provides a nice kabel table to summarize a continuous variable supplied in y}
\usage{
contkabelmat(y, mylab, decimals = 1)
}
\description{
Provides a nice kabel table to summarize a continuous variable supplied in y
}
