% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contkabelmat2.R
\name{contkabelmat2}
\alias{contkabelmat2}
\title{Not needed likely as the calcs aren't so ownerous that we cant just use contkabelmat}
\usage{
contkabelmat2(y, mylab, decimals = 1)
}
\description{
Not needed likely as the calcs aren't so ownerous that we cant just use contkabelmat
}
