% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conttable.R
\name{conttable}
\alias{conttable}
\title{returns a data.frame of summary statistics on continuous vector y that is good for exporting/using elsewhere}
\usage{
conttable(y, mydigits = 1)
}
\description{
returns a data.frame of summary statistics on continuous vector y that is good for exporting/using elsewhere
}
