% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conttablehm.R
\name{conttablehm}
\alias{conttablehm}
\title{returns a data.frame of summary statistics on a time variable y that is in fractional hours and converts to HH:MM; good for exporting/usine elsewhere}
\usage{
conttablehm(y, mydigits = 1)
}
\description{
returns a data.frame of summary statistics on a time variable y that is in fractional hours and converts to HH:MM; good for exporting/usine elsewhere
}
