% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contvarsum.R
\name{contvarsum}
\alias{contvarsum}
\title{provides a nice summary for a continuous variable, with optional boxplot, histogram}
\usage{
contvarsum(mat, idcol, varcol, doplot = T, myuni = T, check5 = F, print5 = T)
}
\description{
provides a nice summary for a continuous variable, with optional boxplot, histogram
}
