% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmcol.R
\name{getmcol}
\alias{getmcol}
\title{determine the (multiple) column number that has the column names}
\usage{
getmcol(mymat, colvec)
}
\description{
determine the (multiple) column number that has the column names
}
