% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time2eventvarsum.R
\name{time2eventvarsum}
\alias{time2eventvarsum}
\title{provides a nice summary for time to event variable, with optional Kaplan-Meier curve}
\usage{
time2eventvarsum(mat, idcol, varcol, cencol, doplot = T)
}
\description{
provides a nice summary for time to event variable, with optional Kaplan-Meier curve
}
